+++ CHANGES.md
@@ -40,6 +40,7 @@
    - Top navigation bar is now displayed on error screens
    - Fixed a problem with icons in calendar appointment tooltips
    - Improved parsing of e-mail message subjects
+   - Ticket owners are now unchanged when a bulk ticket action gets cancelled
  - Fixed a problem with diagnostics information being displayed in rich text editor popups
  - Updated jQuery UI to version 1.13.1